StageBlockの頂点確認用ボックスと中心座標の確認の非表示

+++ TGS2024/FragileWall.h
@@ -23,7 +23,7 @@ private:
 	void BrokenAnimation();				// 壊れるアニメーション
 
 public:
-	void Damege(int damege);            // 被ダメージ処理
+	void Damege(float damege);            // 被ダメージ処理
 
 	bool GetDeleteFlg() const { return delete_flg; }
